[#3098] feat: revise EventProcessors - change inheritance to composition
