+++ messaging/src/main/java/org/axonframework/eventhandling/AbstractEventProcessor.java
@@ -25,20 +25,13 @@ import org.axonframework.messaging.MessageStream;
 import org.axonframework.messaging.unitofwork.ProcessingContext;
 import org.axonframework.messaging.unitofwork.UnitOfWork;
 import org.axonframework.monitoring.MessageMonitor;
-import org.axonframework.monitoring.NoOpMessageMonitor;
-import org.axonframework.tracing.NoOpSpanFactory;
-import org.axonframework.tracing.SpanFactory;
 
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 import java.util.concurrent.CopyOnWriteArrayList;
-
-import static org.axonframework.common.BuilderUtils.assertNonNull;
-import static org.axonframework.common.BuilderUtils.assertThat;
 
 /**
  * Abstract implementation of an {@link EventProcessor}. Before processing of a batch of messages this implementation
@@ -66,14 +59,14 @@ public abstract class AbstractEventProcessor implements EventProcessor {
     protected final EventProcessorSpanFactory spanFactory;
 
     /**
-     * Instantiate a {@link AbstractEventProcessor} based on the fields contained in the {@link Builder}.
+     * Instantiate a {@link AbstractEventProcessor} based on the fields contained in the {@link EventProcessorBuilder}.
      * <p>
      * Will assert that the Event Processor {@code name}, {@link EventHandlerInvoker} and {@link ErrorHandler} are not
      * {@code null}, and will throw an {@link AxonConfigurationException} if any of them is {@code null}.
      *
-     * @param builder the {@link Builder} used to instantiate a {@link AbstractEventProcessor} instance
+     * @param builder the {@link EventProcessorBuilder} used to instantiate a {@link AbstractEventProcessor} instance
      */
-    protected AbstractEventProcessor(Builder builder) {
+    protected AbstractEventProcessor(EventProcessorBuilder builder) {
         builder.validate();
         this.name = builder.name;
         this.eventHandlerInvoker = builder.eventHandlerInvoker;
@@ -249,126 +242,5 @@ public abstract class AbstractEventProcessor implements EventProcessor {
      */
     protected void reportIgnored(EventMessage<?> eventMessage) {
         messageMonitor.onMessageIngested(eventMessage).reportIgnored();
-    }
-
-    /**
-     * Abstract Builder class to instantiate a {@link AbstractEventProcessor}.
-     * <p>
-     * The {@link ErrorHandler} is defaulted to a {@link PropagatingErrorHandler}, the {@link MessageMonitor} defaults
-     * to a {@link NoOpMessageMonitor} and the {@link EventProcessorSpanFactory} defaults to
-     * {@link DefaultEventProcessorSpanFactory} backed by a {@link NoOpSpanFactory}. The Event Processor {@code name}
-     * and {@link EventHandlerInvoker} are <b>hard requirements</b> and as such should be provided.
-     */
-    public abstract static class Builder {
-
-        protected String name;
-        protected EventHandlerInvoker eventHandlerInvoker;
-        protected ErrorHandler errorHandler = PropagatingErrorHandler.INSTANCE;
-        protected MessageMonitor<? super EventMessage<?>> messageMonitor = NoOpMessageMonitor.INSTANCE;
-        protected EventProcessorSpanFactory spanFactory = DefaultEventProcessorSpanFactory.builder()
-                                                                                        .spanFactory(NoOpSpanFactory.INSTANCE)
-                                                                                        .build();
-
-        /**
-         * Sets the {@code name} of this {@link EventProcessor} implementation.
-         *
-         * @param name a {@link String} defining this {@link EventProcessor} implementation
-         * @return the current Builder instance, for fluent interfacing
-         */
-        public Builder name(@Nonnull String name) {
-            assertEventProcessorName(name, "The EventProcessor name may not be null or empty");
-            this.name = name;
-            return this;
-        }
-
-        /**
-         * Sets the {@link EventHandlerInvoker} which will handle all the individual {@link EventMessage}s.
-         *
-         * @param eventHandlerInvoker the {@link EventHandlerInvoker} which will handle all the individual
-         *                            {@link EventMessage}s
-         * @return the current Builder instance, for fluent interfacing
-         */
-        public Builder eventHandlerInvoker(@Nonnull EventHandlerInvoker eventHandlerInvoker) {
-            assertNonNull(eventHandlerInvoker, "EventHandlerInvoker may not be null");
-            this.eventHandlerInvoker = eventHandlerInvoker;
-            return this;
-        }
-
-        /**
-         * Sets the {@link ErrorHandler} invoked when an {@link UnitOfWork} throws an exception during processing.
-         * Defaults to a {@link PropagatingErrorHandler}.
-         *
-         * @param errorHandler the {@link ErrorHandler} invoked when an {@link UnitOfWork} throws an exception during
-         *                     processing
-         * @return the current Builder instance, for fluent interfacing
-         */
-        public Builder errorHandler(@Nonnull ErrorHandler errorHandler) {
-            assertNonNull(errorHandler, "ErrorHandler may not be null");
-            this.errorHandler = errorHandler;
-            return this;
-        }
-
-        /**
-         * Sets the {@link MessageMonitor} to monitor {@link EventMessage}s before and after they're processed. Defaults
-         * to a {@link NoOpMessageMonitor}.
-         *
-         * @param messageMonitor a {@link MessageMonitor} to monitor {@link EventMessage}s before and after they're
-         *                       processed
-         * @return the current Builder instance, for fluent interfacing
-         */
-        public Builder messageMonitor(@Nonnull MessageMonitor<? super EventMessage<?>> messageMonitor) {
-            assertNonNull(messageMonitor, "MessageMonitor may not be null");
-            this.messageMonitor = messageMonitor;
-            return this;
-        }
-
-        /**
-         * Sets the {@link EventProcessorSpanFactory} implementation to use for providing tracing capabilities. Defaults
-         * to a {@link DefaultEventProcessorSpanFactory} backed by a {@link NoOpSpanFactory} by default, which provides
-         * no tracing capabilities.
-         *
-         * @param spanFactory The {@link SpanFactory} implementation
-         * @return The current Builder instance, for fluent interfacing.
-         */
-        public Builder spanFactory(@Nonnull EventProcessorSpanFactory spanFactory) {
-            assertNonNull(spanFactory, "SpanFactory may not be null");
-            this.spanFactory = spanFactory;
-            return this;
-        }
-
-        /**
-         * Validates whether the fields contained in this Builder are set accordingly.
-         *
-         * @throws AxonConfigurationException if one field is asserted to be incorrect according to the Builder's
-         *                                    specifications
-         */
-        protected void validate() throws AxonConfigurationException {
-            assertEventProcessorName(name, "The EventProcessor name is a hard requirement and should be provided");
-            assertNonNull(eventHandlerInvoker, "The EventHandlerInvoker is a hard requirement and should be provided");
-        }
-
-        private void assertEventProcessorName(String eventProcessorName, String exceptionMessage) {
-            assertThat(eventProcessorName, name -> Objects.nonNull(name) && !"".equals(name), exceptionMessage);
-        }
-
-        public String name() {
-            return name;
-        }
-
-        public EventHandlerInvoker eventHandlerInvoker() {
-            return eventHandlerInvoker;
-        }
-
-        public ErrorHandler errorHandler() {
-            return errorHandler;
-        }
-
-        public MessageMonitor<? super EventMessage<?>> messageMonitor() {
-            return messageMonitor;
-        }
-
-        public EventProcessorSpanFactory spanFactory() {
-            return spanFactory;
-        }
     }
 }
